pyproject.toml
[tool.poetry]
name = "compliance-rag"
version = "0.1.0"
description = ""
authors = ["ViratCh04 <105713825+ViratCh04@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
isort = "^5.13.2"
black = "^24.8.0"
cohere = "^5.9.2"
pydantic = "^2.9.2"
qdrant-client = "^1.11.2"
openai = "^1.47.0"
langchain = "^0.3.0"
langchain-community = "^0.3.0"
langchain-cohere = "^0.3.0"
langchain-openai = "^0.2.0"
langgraph = "^0.2.28"
langchain-qdrant = "^0.1.4"
python-dotenv = "^1.0.1"
pypdf = "^5.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
